completed week 3

+++ Week 3/Sheet 5/p5p3.py
@@ -1,6 +1,6 @@
 # p5p3 - Program checks whether a number is positive, negative or equal to zero
 
-user_num = int(input("Please type an integer: "))
+user_num = float(input("Please type an integer: "))
 
 if user_num > 0:
     print(user_num, "is positive")
